# swagger.json
{
  "consumes": [
    "application/json",
    "application/xml"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "the purpose of this application is to provide an application\nthat is using plain go code to define an API\n\nThis should demonstrate all the possible comment annotations\nthat are available to turn go code into a fully compliant swagger 2.0 spec",
    "title": "SmallShop API.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "contact": {
      "name": "John Doe",
      "url": "http://john.doe.com",
      "email": "john.doe@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/v2",
  "paths": {
    "/foobar": {
      "post": {
        "tags": [
          "foobar-tag"
        ],
        "summary": "Foobar does some amazing stuff.",
        "operationId": "idOfFoobarEndpoint",
        "parameters": [
          {
            "description": "This text will appear as description of your request body.",
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/FooBarRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/foobarResponse"
          }
        }
      }
    },
    "/health": {
      "get": {
        "description": "Returns 204 without content",
        "tags": [
          "general"
        ],
        "summary": "Check if the service is running.",
        "operationId": "Health",
        "responses": {
          "204": {
            "description": "Service available"
          }
        }
      }
    },
    "/oauth/token": {
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Authenticate and authorise with Windows credentials.",
        "operationId": "GetToken",
        "parameters": [
          {
            "type": "string",
            "default": "password",
            "description": "Grant type",
            "name": "grant_type",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "description": "domain\\username",
            "name": "username",
            "in": "formData",
            "required": true
          },
          {
            "type": "string",
            "format": "password",
            "description": "Your password",
            "name": "password",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "definitions": {
    "FooBarRequest": {
      "type": "object",
      "title": "FooBarRequest represents body of FooBar request.",
      "properties": {
        "bar": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "x-go-name": "Bar"
        },
        "foo": {
          "type": "string",
          "x-go-name": "Foo"
        }
      },
      "x-go-package": "github.com/pdrum/swagger-automation/api"
    },
    "FooBarResponse": {
      "type": "object",
      "title": "FooBarResponse represents body of FooBar response.",
      "properties": {
        "baz": {
          "type": "object",
          "properties": {
            "prop": {
              "type": "string",
              "x-go-name": "Prop"
            }
          },
          "x-go-name": "Baz"
        }
      },
      "x-go-package": "github.com/pdrum/swagger-automation/api"
    }
  },
  "responses": {
    "foobarResponse": {
      "description": "This text will appear as description of your response body.",
      "schema": {
        "$ref": "#/definitions/FooBarResponse"
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "KEY",
      "in": "header"
    },
    "oauth2": {
      "type": "oauth2",
      "in": "header",
      "flow": "accessCode",
      "authorizationUrl": "/oauth2/auth",
      "tokenUrl": "/oauth2/token",
      "scopes": {
        "bar": "foo"
      }
    }
  },
  "security": [
    {
      "api_key": []
    }
  ],
  "x-meta-array": [
    "value1",
    "value2"
  ],
  "x-meta-array-obj": [
    {
      "name": "obj",
      "value": "field"
    }
  ],
  "x-meta-value": "value"
}